+++ src/index.js
@@ -1,0 +1,3 @@
+import CloudentityWebAuth from './CloudentityWebAuth';
+
+module.exports =  CloudentityWebAuth; // not 'export default' because of issues with webpack's UMD export