+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [Unreleased]
+### Added
+- initial client (basic OAuth implicit flow support) - WIP
+- basic options validation
+- basic documentation - WIP
+- build config
+
 ## [Project init]
 ### Added
 - initial project files